close to finishing mesh class
Updated the vertex array to be more flexible with buffer objects and it
now allows for multiple buffer objects. Buffer objects for indices and
vertex buffers can now be switched and bound to the vertex array more
easily. Still uncertain on the design for the renderer class, as well as
whether the mesh and/or vertex array classes should retain information
about data formatting for easier rendering, and if so how it should be
retained.

Changed Triangle class to Face class.

Made Edge and HalfEdge separate classes instead of nested classes of
face.

Updated the hashCode functions for multple classes that use them. Now
each class uses the Arrays.HashCode function to produce its hashCode.