Finalized the vertex array and mesh classes
Created a IndexBuffer class for handling the data for indices as well as
maintaining meta data about the rendering of the index.

Also introduced a Transform class to handle object transforms. The class
retains an orientation, position, and scaling.

Created a RenderMode enumto restrict render specifications with the
IndexBuffer.. The enum specifies all the GLenum rendering modes
available. Also created an IndexType to specify how an IndexBuffer will
store and transfer its data, either as bytes, shorts, or intergers.