Nearly finished with PNG parser
PNG parser can handle all bit depths and color types. Issues arose with
the 16-bit depths. The problem could have been an issue with texture
internal formatting or possibly HDR effecting the rendering of the
texture. Currently the issue is resolved through reduction of values to
8-bit until the issue can be found and resolved. This will be the
handling method for greyscale and truecolor. Greyscale with alpha still
needs to be implemented for the above method.

Additionally added unit tests for image parsing with PNG being mostly
filled out. Still need to add a few test cases mostly with adam 7
interlacing and with additional chunk data that is optional to parse.