updated diagrams and condensed ImageLoader
Updated the documentation diagrams to reflect the current state of the
image loading classes. Added comments to the PNGDecoder and to
ImageParser.

Condensed the ImageLoader class to use only the load functions and to
make use of the generality of the ImageParser base class. This means
that now all that needs to be done is to add another conditional to the
portion that generates the decoders to decide which decoder to use based
on the image file extension. This makes the code much more maintainable
for future expansion.