Adjusted the light classes to bind to specialized light structs in shaders
however the results of this are unpleasant so I am reverting it back
however I am using the github versioning to retain the current state for
potential future use.

Additionally updated the vectors normalize functions to return the
vector post normalization instead of the length.

Added coloring to the arrow class and setup the arrow and
TransformControl to set colors for rendering themselves