collisions with rays and AABB's working
The collision detection between AABB's and Rays is functional, the next
step is to implement a system for projecting a ray into the scene based
on information stored in the camera. This means the camera class will
require modifications and improvements to create a more streamlined
system for generating these rays.