Fixed bugs with the convex hull generation
Additionally updated the hull gen class to better view the construction
of the hull. The hull generation still hasn't been set up to handle
co-planar geometry such as discs.

In addition the Geometry class now updates a list of the farthest
vertices of the geometry each time a new vertex is updated. This allows
the hull generation code to build the initial hull tetrahedra faster.
And this can provide a way to generate mesh dimension, and create AABB.

The HalfEdge class had a new function added that checks whether a given
HalfEdge is a valid opposite HalfEdge to act as an opposite to the edge
pair.