Some fixes and preventative code added
Uniform had some error preventing code added to it to maintain buffer
expectations when setting a uniform value with less data than is
required such as using Vec2 to set the first 2 values of a Vec4.
Additionally fixed Ray generation with the camera function. Finally
current experimenting with using the support function to integrate ray
collision detection with the GJK algorithm function to simplify
collision detection.