Working on more collision detection
Doing some integration work with the new Entity class that is meant to
unify Renderable meshes and CollisionMeshes. Additionally working on
more collision detection that needs to be done, in particular Ray convex
hull collision detection. Further collisino detection will be added,
most likely capsule collision meshes, and chained collision meshes, this
will allow for more accurate collision detection with non convex meshes
like the torus. This is necessary for completion of the Transform gozmos
that will be used for easier transformations of in world entities for
render testing.

From there more work will be done to improve the Renderer to make it
more adaptive and configurable. Currently it has a somewhat static
design to it that is not favorable to more dynamic scenes  or rendering
techniques.