Ray intersect with convex hull 2D should be complete
Intersecting a ray with a 2D convex hull should complete, final step is
to doing some extensive testing with a small application.

Additionally added some functionality to the transform class to make
transformation of 3D vectors more stream lined, which is where the
refactor of some of the collision meshes comes in.