Updated collision detection algorithm
Updated the collision detection algorithm for comparing if a Ray and 2D
Convex hull are intersecting. I changed the code for determining
intersection when the ray and hull are parallel to utilize the
parametric line clipping algorithm. The old code had faults in it that
could fail under certain circumstances. The new code should be able to
handle those cases, and will be more thoroughly tested in the future
with unit testing.

Next intersections with Ray and 3D hulls. Currently looking into using
3D line clipping to accomplish this, however need to research some more
and consider my options.