Fixed a bug with the parametric line clipping algorithm
The algorithm was using a scale of 0 to 1 to determine the depth along
the line to use for entering and leaving the hull. But this didn't
translate as well into my classes since the direction vector used in the
original code had the length of the ray as a part of it, making the 0 to
1 calculations correlate to the ray in world space with teh length
already considered. However my ray class returned a normalized direction
vector meaning 0 to 1 depth didn't account for the full ray length.