Main focus so far has been on the transformation gizmos
The main focus of the work done so far is to get the transformation
gizmos working so that scenes can be more interact. It will make it
easier to adjust objects to test changing environments, as opposed to
binding keys to perform different transformations.

One of the bigger changes is that the mesh class no longer is a
SpatialAsset. This means that the Mesh class will no longer maintain its
own instance of a Transform object. Instead I decided that it would be
better to shift that focus to the Entity class, which is a SpatialAsset.
This also helps to reduce ambiguity when using Mesh's in the Entity
class, since it reduces external asset manipulation ona  shared
functionality, and removes prioritization problems with the Transforms.

Smaller changes made were to the listener classes for input events. The
main functions to for taking in input had "on" appended to their names
to clear up any confusion with the functionality of the methods.

Additionally updated any classes affected by the change made to the Mesh
class no longer being a SpatialAsset.