Mainly working on transformation gizmos
Mainly focusing on getting the transformation gizmos working. Currently
have the basic sphere interaction for the translation gizmo working.

Fixed some bugs with the matrix classes. moved the axis constant vectors
from VecUtil to the Transform class, and made it so the Camera class now
takes the width and height as parameters.

Currently trying to work out a bug in the collision detection code for
Rays and 3D collision meshes. There seems to be some type of problem
when rotating the mesh any degrees around the x axis, causing detection
errors.