Finshed main functionality of the Translation gizmo
The primary work need for the TranslationGizmo class has been finished.
Now objects can be translated using the tool in real time in the 3D
environment without the need for key bindings.

Currently working on the Scale gizmo class. Currently working
effectively for what it should do. However more testing needs to be done
to adjust values to help prevent rapid scaling issues based on camera
distances.

After this class is finished the last gizmo I will be working on it the
rotation gizmo. This one will prove to be more complicated due to using
torus shapes as the control wheels for the rotation. These will require
hull decomposition to function properly.