Rotation gizmo working for the most part
The rotation gizmo is functionally working right now, there are a few
issues though. Currently there appears to be a bug in which the object
the gizmo is bound to will disappear while the gizmo is interacting with
it. This might hav something to do with a bad matrix when rotations
accumulate in the quaternion class.

Additionally the convex hull used in interacting with the control wheels
for the rotation gizmo are basic and don't account for the hole in the
torus. This can be solved when I explore more into convex hull
decomposition which should be soon.