Fixed errors with the scaling gizmo
Fixed a few bugs in the scaling gizmo. The first bug was happening when
the gizmo attempted to undo previous rotations applied to it to orient
itself to a new binding. The second bug was in the way the scaling
calculations were being done resulting in scale values accumulating in
fractional amounts in the mesh. This would make the scaling action
permanent and inconsistant.