Updated the vertex array class
Updated the vertex array class to create and store it's own vertex
buffers and index buffers. This is meant to help reduce needed code in
the user classes and provide a mode predictable usage of the vertex
array class. Additionally the vertex array class already maintained
constant pointers to the vbos and ibos, making the mesh classes
retaining of pointers redundant.

Additionally modified the Bexier path class, improving it and adding
more fucntionality. The end goal of which is create a class more robust
at handling Bexier curves for use in constructing renderable text.  from
truetype fonts and other similar font formats.