Fixed a bug with the cylinder class
Additionally changed the way the Bezier path class gets the derivative
and length. Previously these were being called everytime a control point
was added or in other various methods. This became a problem as curves
were being constructed, causing slow downs in font loading. Instead a
flag is used to inidicate these values need to be updated and are
updated when necessary, such as when using the length or derivative.

Also continued work on the font loading, still need to work with
composite glyphs and load those, after which rendering testing can
begin.