Continuing work on font loading
At this point I may consider alternative methods to load the fonts
instead of directly loadnig the file data and attempting to render the
curves. This will be taken into consideration going forward.

Additionally added some simple functionality to use a configuration file
in the form of a json file. This will allow easy changes to the inputs
for certain functions for testing purposes, reducing the need to
recompile code for simple test changes.