Added configuration file for easy settings changes
Also worked on recursive generation of bezier curves. This proved
difficult to make adaptive to the curve and efficient and will be
scrapping it. However I am adding the code to the version control for
future reference if I choose to return to it.

Also added a configuration manager to be able to store settings in a
configuration file to more readily change code parameters where
necessary.