Made changes to the ResourceManager
ResourceManager is now a generic templated type, since the other
managers weren't providing any other benefits other than converting the
type for other functions generics solved this.

Also moved shader program definitions and management to a new class that
loads the shader programs from the configuration file on loading the
system. This allows for much easier loading and grouping of the shader
programs using the configuration file. This is just one step in my
attempt to improve the code for the renderer classes and make them more
robust.